backend/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
APScheduler==3.8.1
astroid==2.8.6
attrs==21.4.0
bcrypt==4.0.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
distlib==0.3.6
dnspython==2.1.0
email-validator==1.1.3
filelock==3.8.0
Flask==2.0.2
Flask-Admin==1.5.8
Flask-APScheduler==1.12.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
lazy-object-proxy==1.6.0
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
pipenv==2022.9.8
platformdirs==2.4.0
protobuf==3.19.4
psycopg2-binary==2.9.3
PyJWT==2.3.0
pypi==2.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
